feat: add student controller
